Improved MCTS performance

+++ coro.cpp
@@ -15,7 +15,7 @@ namespace coro {
         //do work
         volatile unsigned long x = 0;
         for(int i = 0; i < 1; i++) {
-            x = x + std::chrono::system_clock::now().time_since_epoch().count();
+            x =  x + (unsigned long) std::chrono::system_clock::now().time_since_epoch().count();
             //std::cout << x << std::endl;
         }
         cnt++;
